requirements.txt
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
tqdm
Pillow
sentencepiece
protobuf
scipy
matplotlib
seaborn